+++ src/io/mod.rs
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
+pub mod buttons;
 pub mod console;
 pub mod ow_temp;
 pub mod pps;
@@ -6,3 +7,5 @@ pub mod rpm;
 pub mod shared_i2c;
 #[cfg(feature = "serial-cmd")]
 pub mod serial_cmd;
+pub mod touch_buttons;
+pub mod watchdog;